requirements.txt
Flask>=2.2.2, <2.2.3
flask-socketio>=5.3.2, <5.3.3
eventlet>=0.33.2, <0.33.3
flake8>=6.0.0, <6.0.1
pytest>=7.2.0, <7.2.1
psycopg2-binary>=2.9.5, <2.9.6